refactor: split event detail page
